+++ src/components/util/mixins/plugin/print.js
@@ -1,0 +1,2 @@
+// 参考： https://printjs.crabbly.com/
+export * as print from "print-js"
