+++ src/components/util/mixins/print.js
@@ -1,0 +1,14 @@
+import { print } from './plugin/print'
+
+export default {
+  computed: {
+    opts: function() {
+      return null
+    }
+  },
+  methods: {
+    print: function() {
+      print(this.opts);
+    },
+  },
+}
